requirements.txt
arrow==0.15.5
asgiref==3.2.7
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-extensions==2.2.9
django-filter==2.2.0
django-heroku==0.3.1
django-nose==1.4.6
django-redis==4.11.0
django-tinymce==2.8.0
djangorestframework==3.11.0
form==0.0.1
googlemaps==4.2.0
idna==2.9
inflection==0.3.1
Inflector==3.0.1
jsonfield==3.1.0
kafka-python==2.0.1
nose==1.3.7
Pillow==7.0.0
psycopg2==2.8.5
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2019.3
querystring-parser==1.2.4
redis==3.4.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
whitenoise==5.0.1
gunicorn